--- test_scripts/Polices/Validation_of_PolicyTables/011_ATF_preloaded_pt_validation_rules_for_required_parameters.lua
@@ -1,114 +1,0 @@
----------------------------------------------------------------------------------------------
--- Requirement summary:
--- [Policies]: PreloadedPolicyTable: Validation rules for required parameters
---
--- Description:
--- Behavior of SDL during start SDL in case when PreloadedPT has no required parameters
--- 1. Used preconditions:
--- Delete files and policy table from previous ignition cycle if any
--- Do not start default SDL
--- 2. Performed steps:
--- Create PreloadedPolicyTable file without one required parameter
--- Start SDL
---
--- Expected result:
--- SDL is shutdown
----------------------------------------------------------------------------------------------
---[[ General configuration parameters ]]
-Test = require('connecttest')
-local config = require('config')
-config.defaultProtocolVersion = 2
-
---[[ Required Shared libraries ]]
-local commonFunctions = require ('user_modules/shared_testcases/commonFunctions')
-local commonSteps = require('user_modules/shared_testcases/commonSteps')
-local json = require("modules/json")
-local SDL = require('modules/SDL')
-
---[[ Local Variables ]]
-local PRELOADED_PT_FILE_NAME = "sdl_preloaded_pt.json"
-local TEST_DATA = {
-  ["1234"] = {
-    [123] = "http://cloud.ford.com/global",
-    keep_context = false,
-    steal_focus = false,
-    -- priority = "NONE", -- removed required parameter
-    default_hmi = "NONE",
-    groups = {"BaseBeforeDataConsent"}
-  },
-  super = {
-    keep_context = false,
-    steal_focus = false,
-    priority = "NONE",
-    default_hmi = "NONE",
-    groups = {"BaseBeforeDataConsent"}
-  }
-}
-
---[[ Local Functions ]]
-function Test.backupPreloadedPT()
-  os.execute("cp " .. config.pathToSDL .. PRELOADED_PT_FILE_NAME .. ' ' .. config.pathToSDL .. "backup_" .. PRELOADED_PT_FILE_NAME)
-end
-
-function Test.restorePreloadedPT()
-  os.execute("mv " .. config.pathToSDL .. "backup_" .. PRELOADED_PT_FILE_NAME .. " " .. config.pathToSDL .. PRELOADED_PT_FILE_NAME)
-end
-
-function Test.corruptPreloadedPT(changed_parameters)
-  local pathToFile = config.pathToSDL .. PRELOADED_PT_FILE_NAME
-
-  local file = io.open(pathToFile, "r")
-  local json_data = file:read("*a")
-  file:close()
-
-  local data = json.decode(json_data)
-  if data then
-    for key, value in pairs(changed_parameters) do
-      data.policy_table.app_policies[key] = value
-    end
-  end
-
-  local dataToWrite = json.encode(data)
-  file = io.open(pathToFile, "w")
-  file:write(dataToWrite)
-  file:close()
-end
-
-function Test.checkSdl()
-  EXPECT_HMINOTIFICATION("BasicCommunication.OnSDLClose",{}):Times(1)
-  local status = SDL:CheckStatusSDL()
-  if status == SDL.RUNNING then
-    commonFunctions:userPrint(31, "Test failed: SDL running without required parameters in preloaded policy table")
-    return false
-  end
-  return true
-end
-
---[[ Preconditions ]]
-function Test:Precondition_StopSdl()
-  StopSDL(self)
-end
-
-function Test:Precondition()
-  commonSteps:DeletePolicyTable()
-  self.backupPreloadedPT()
-  self.corruptPreloadedPT(TEST_DATA)
-end
-
---[[ Test ]]
-function Test:Test_StartSdl()
-  StartSDL(config.pathToSDL, true, self)
-end
-
-function Test:Test()
-  os.execute("sleep 3")
-  self.checkSdl()
-end
-
---[[ Postconditions ]]
-function Test:Postconditions()
-  self.restorePreloadedPT()
-end
-
-commonFunctions:SDLForceStop()
-return Test
